requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.1.3
Django==5.0.6
django-ninja==1.1.0
djangorestframework==3.15.1
ecdsa==0.19.0
jose==1.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
